Only stop index container if one is actually running (#70)
So far, `qlever index` always called `docker rm -f ...` or `podman rm -f ...` before starting a new index build. However, old versions of podman return an error message when no container with the specified name is running. This is now fixed by checking explicitly whether a container is running and only if there is, trying to stop and remove it. Fixes #66.

+++ src/qlever/commands/index.py
@@ -119,11 +119,14 @@ class IndexCommand(QleverCommand):
         # Remove already existing container.
         if args.system in Containerize.supported_systems() \
                 and args.overwrite_existing:
-            try:
-                run_command(f"{args.system} rm -f {args.index_container}")
-            except Exception as e:
-                log.error(f"Removing existing container failed: {e}")
-                return False
+            if Containerize.is_running(args.system, args.index_container):
+                log.info("Another index process is running, trying to stop it ...")
+                log.info("")
+                try:
+                    run_command(f"{args.system} rm -f {args.index_container}")
+                except Exception as e:
+                    log.error(f"Removing existing container failed: {e}")
+                    return False
 
         # Write settings.json file.
         try:
